+++ source/winBindings/cfgmgr32.py
@@ -10,6 +10,9 @@ from ctypes.wintypes import DWORD, ULONG
 
 dll = windll.cfgmgr32
 
+CR_SUCCESS = 0
+MAX_DEVICE_ID_LEN = 200
+
 CM_Get_Device_ID = dll.CM_Get_Device_IDW
 """
 Retrieves the device instance ID for a specified device instance on the local machine.